rename: 검증 관련 패키지 이름을 validation으로 통일, 검증 관련 클래스의 접미사를 validator로 통일
- 해당 변경은 meeting, participation, profile 패키지에만 적용
