+++ src/main/java/com/momo/profile/service/ProfileService.java
@@ -5,8 +5,8 @@ import com.momo.profile.dto.ProfileCreateRequest;
 import com.momo.profile.dto.ProfileCreateResponse;
 import com.momo.profile.entity.Profile;
 import com.momo.profile.repository.ProfileRepository;
-import com.momo.profile.validate.ProfileRequiredValueValidator;
-import com.momo.profile.validate.ProfileValidator;
+import com.momo.profile.validation.ProfileRequiredValueValidator;
+import com.momo.profile.validation.ProfileValidator;
 import com.momo.user.entity.User;
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
